Rename as templates
