+++ lib/middleware/template-gambit.js
