+++ lib/middleware/template-michael.js
